v1/src/main/webapp/app/entities/data-forest/delete/data-forest-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DataForestService } from '../service/data-forest.service';

import { DataForestDeleteDialogComponent } from './data-forest-delete-dialog.component';

describe('DataForest Management Delete Component', () => {
  let comp: DataForestDeleteDialogComponent;
  let fixture: ComponentFixture<DataForestDeleteDialogComponent>;
  let service: DataForestService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, DataForestDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(DataForestDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DataForestDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DataForestService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
